update card manager

+++ card-manager/src/models/EquipmentCard.java
@@ -3,7 +3,7 @@ package models;
 public class EquipmentCard extends Card {
     private int bonusPower;
     private int bonusHealth;
-    public EquipmentCard(String id, String cardName, int cost, String region, long price, byte quantity, String cardText, String keyword, int bonusPower, int bonusHealth) {
+    public EquipmentCard(String id, String cardName, int cost, String region, double price, byte quantity, String cardText, String keyword, int bonusPower, int bonusHealth) {
         super(id, cardName, cost, region, price, quantity, cardText, keyword);
         this.bonusPower = bonusPower;
         this.bonusHealth = bonusHealth;
