+++ card-manager/src/models/SpellCard.java
@@ -3,7 +3,7 @@ package models;
 public class SpellCard extends Card{
     private int speed; // 1-4, slow, fast, focus, burst
 
-    public SpellCard(String id, String cardName, int cost, String region,long price, byte quantity, String cardText, String keyword, int speed) {
+    public SpellCard(String id, String cardName, int cost, String region,double price, byte quantity, String cardText, String keyword, int speed) {
         super(id, cardName, cost, region, price, quantity, cardText, keyword);
         this.speed = speed;
     }
