+++ card-manager/src/models/UnitCard.java
@@ -5,7 +5,7 @@ public class UnitCard extends Card{
     private int health;
     private boolean isChampion;
 
-    public UnitCard(String id, String cardName, int cost, String region,long price, byte quantity, String cardText, String keyword, int power, int health, boolean isChampion) {
+    public UnitCard(String id, String cardName, int cost, String region,double price, byte quantity, String cardText, String keyword, int power, int health, boolean isChampion) {
         super(id,cardName,cost,region,price,quantity, cardText, keyword);
         this.power = power;
         this.health = health;
